move `FPMathMode` to frontend
